requirements.txt
openai>=1.30.0
chromadb>=0.5.0
python-dotenv>=1.0.0
rich>=13.7.0
tiktoken>=0.7.0
pytest>=8.0.0
